Global Servers + Disc Auth
This update reworked the our entire discord authentication system to allow us to expand our use for discord in the future for things such as JFR's. The update also made it easier to change things in the future and fetch user data. This update also finished the UI and backend for the global servers feature that will be released.